Created Analysis for usage of Array elements
The analysis is based on Cousot et al. "A parametric segmentation functor for fully automatic and scalable array content analysis" (2011) and the Master Thesis of Jan Haltermann. The current version is a prototypic implementation.
It contains two main components: On the one hand, there is the implementation of the Array Segmentation domain, containing everything needed to lift an analysis for simple variables to arrays. On the other hand, we have a simple analysis to detect usage of array elements, showing how to use the array segmentation domain.
It is assumed, that no “Array-Index-out-of-bound” errors occur in the analysed program.

Known issues that need to be fixed and planned extentions:

- The analysis may not work correctly if there are branches in loops, where the number of statements in the branches differs
- Currently, we are assuming that the variable used to access the array(‘i’) is always bounded above by the length of the array. Moreover, we assume that the size of the array does not overflow, hence we can assume that ‘i+1-1 = I’ holds (same assumption as in original paper).
- The path formula is currently computed using a slight modified implementation from Manuel Tows. We plan to use the predicate analysis to compute the path formula.